10/24 백준 문제풀이

+++ Java/src/baekJoon/Ex19_1000.java
@@ -7,11 +7,10 @@ public class Ex19_1000 {
 	public static void main(String[] args) {
 		Scanner scn = new Scanner(System.in);
 		
-		int a = scn.nextInt();
-		int b = scn.nextInt();
+		double a = scn.nextDouble();
+		double b = scn.nextDouble();
 		
-		System.out.println((float)a/b);
+		System.out.println(a/b);
 
 	}
-
 }
